cg-nat: add FP filling
fp-filling is able to provide cgnat mempool filing. So we want to add
the filling panel to the cg-nat conf. But, with cg-nat stats, we are
already able to display history of number entries in mempool.

So we don't need to have base-filling-history panel in cgnat dashboard.
We need to only have base-filling-pie panel.

Unfortunately, this one can only be repeated in different rows. But,
only one pie can not full a row. You need to have the pie and the
history diagram to fill a row.

In fact, what we want it's to have multiple pies in the same row for the
cg-nat. To do that, we duplicate base-filling-pie panel into a new one
panel base-filling-pie-repeat. This one has the following extra parameter:
"repeat": "filling".

The title is also changed from 'Current filling' to `$filling`.
Because when the base-filling-pie, `$filling` is used as title of the
row. It can not be the case when we have different filling pies in the
same row.

Signed-off-by: Maxime Leroy <maxime.leroy@6wind.com>