panels/cg-nat: don't display min and max for blocks, cpes
In blocks-cpe panel, we show the number of cpes and blocks entries.
We also show in the legend: minimun and maximun values.

It was useful to know what's the maximun of sessions allocated to
estimated number of sessions supported by the system.

But now we have the pie filling to show that. So we can remove it.

Signed-off-by: Maxime Leroy <maxime.leroy@6wind.com>