upgrade influx to 1.11.8
Let's upgrade, it has been a while.

The permissions work differently with this container, pre-create the
influxdb/data directory and use the local user to launch influx.

The users migrating from a previous version might need to manually set
the owner of the local influxdb/data directory to the local user using
chown -R.

Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>